+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Ensure per-functionl-call amazonica credentials are merged correctly with overriden credentials
+
 ## 1.0.1
 
 - Add `signal.amazonica-aws-sso/default-credentials-provider` as a convenience method for newing up `com.signalai.awsSSOCompatibleDefaultAWSCredentialsProviderChain`